requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
flask
langchain-huggingface
langchain-google-genai
langchain-groq


-e .